+++ pkg/ui/Makefile
@@ -31,23 +31,34 @@ protos:
 generate: $(GOBINDATA_TARGET)
 
 .PHONY: lint
-lint: $(YARN_INSTALLED_TARGET) | protos
+lint: | protos
 	yarn run stylint -c .stylintrc styl
 	yarn run tslint -c tslint.json -p tsconfig.json --type-check
 	@# TODO(benesch): Invoke tslint just once when palantir/tslint#2827 is fixed.
 	yarn run tslint -c tslint.json *.js
 
+# See "DLLs for speedy builds" in the README for details.
+DLLS := dist/protos.dll.js dist/vendor.dll.js protos-manifest.json vendor-manifest.json
+
 .PHONY: test
-test: $(YARN_INSTALLED_TARGET) | protos
+test: $(DLLS)
 	yarn run karma start
 
 .PHONY: test-watch
-test-watch: $(YARN_INSTALLED_TARGET) | protos
+test-watch: $(DLLS)
 	yarn run karma start --no-single-run --auto-watch
 
-$(GOBINDATA_TARGET): $(YARN_INSTALLED_TARGET) $(shell git ls-files | grep -vF $(GOBINDATA_TARGET)) | protos
-	rm -rf dist
-	yarn run webpack -p
+# (Ab)use a pattern rule to teach Make that this one command produces two files.
+# Normally, it would run the recipe twice if dist/FOO.js and FOO-manifest.js
+# were both out-of-date.
+#
+# See: https://stackoverflow.com/a/3077254/1122351
+dist/%.dll.js %-manifest.json: webpack.%.js | protos
+	yarn run webpack -p --config $<
+
+$(GOBINDATA_TARGET): webpack.app.js $(DLLS) $(shell find src)
+	find dist -mindepth 1 -not -name '*.dll.js' -delete
+	yarn run webpack --config $<
 	go-bindata -nometadata -pkg ui -o $@ -prefix dist dist/...
 	# Add comment recognized by reviewable.
 	echo '// GENERATED FILE DO NOT EDIT' >> $@
