+++ pkg/ui/src/js/.gitignore
@@ -1,0 +1,2 @@
+protos.d.ts
+protos.js
