+++ pkg/ui/ui.go
@@ -12,8 +12,5 @@
 // implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-// Package ui embeds into the Cockroach certain data such as web html and
-// stylesheets.
+// Package ui embeds the assets for the web UI into the Cockroach binary.
 package ui
-
-//go:generate make generate
